+++ src/apis/admin/api.ts
@@ -3,7 +3,9 @@
  * @module AdminAPI
  */
 import AvalancheCore from '../../avalanche';
-import { JRPCAPI, RequestResponseData } from '../../utils/types';
+import { JRPCAPI } from '../../common/jrpcapi';
+import { RequestResponseData } from '../../common/apibase';
+
 
 /**
  * Class for interacting with a node's AdminAPI.
