+++ src/apis/keystore/api.ts
@@ -3,7 +3,8 @@
  * @module KeystoreAPI
  */
 import AvalancheCore from '../../avalanche';
-import { JRPCAPI, RequestResponseData } from '../../utils/types';
+import { JRPCAPI } from '../../common/jrpcapi';
+import { RequestResponseData } from '../../common/apibase';
 
 /**
  * Class for interacting with a node API that is using the node's KeystoreAPI.
