+++ src/keychains/keychain.ts
@@ -1,0 +1,206 @@
+/**
+ * @packageDocumentation
+ * @module Common-KeyChain
+ */
+
+import { Buffer } from "buffer/";
+
+/**
+ * Class for representing a private and public keypair in Avalanche. 
+ * All APIs that need key pairs should extend on this class.
+ */
+export class KeyPair {
+    protected pubk:Buffer;
+  
+    protected privk:Buffer;
+  
+    /**
+       * Generates a new keypair.
+       *
+       * @param entropy Optional parameter that may be necessary to produce secure keys
+       */
+    generateKey:(entropy?:Buffer) => void;
+  
+    /**
+       * Imports a private key and generates the appropriate public key.
+       *
+       * @param privk A {@link https://github.com/feross/buffer|Buffer} representing the private key
+       *
+       * @returns true on success, false on failure
+       */
+    importKey:(privk:Buffer) => boolean;
+  
+    /**
+       * Takes a message, signs it, and returns the signature.
+       *
+       * @param msg The message to sign
+       *
+       * @returns A {@link https://github.com/feross/buffer|Buffer} containing the signature
+       */
+    sign:(msg:Buffer) => Buffer;
+  
+    /**
+       * Recovers the public key of a message signer from a message and its associated signature.
+       *
+       * @param msg The message that's signed
+       * @param sig The signature that's signed on the message
+       *
+       * @returns A {@link https://github.com/feross/buffer|Buffer} containing the public
+       * key of the signer
+       */
+    recover:(msg:Buffer, sig:Buffer) => Buffer;
+  
+    /**
+       * Verifies that the private key associated with the provided public key produces the
+       * signature associated with the given message.
+       *
+       * @param msg The message associated with the signature
+       * @param sig The signature of the signed message
+       * @param pubk The public key associated with the message signature
+       *
+       * @returns True on success, false on failure
+       */
+    verify:(msg:Buffer, sig:Buffer, pubk:Buffer) => boolean;
+  
+    /**
+       * Returns a reference to the private key.
+       *
+       * @returns A {@link https://github.com/feross/buffer|Buffer} containing the private key
+       */
+    getPrivateKey = ():Buffer => this.privk;
+  
+    /**
+       * Returns a reference to the public key.
+       *
+       * @returns A {@link https://github.com/feross/buffer|Buffer} containing the public key
+       */
+    getPublicKey = ():Buffer => this.pubk;
+  
+    /**
+       * Returns a string representation of the private key.
+       *
+       * @returns A string representation of the public key
+       */
+    getPrivateKeyString:() => string;
+  
+    /**
+       * Returns the public key.
+       *
+       * @returns A string representation of the public key
+       */
+    getPublicKeyString:() => string;
+  
+    /**
+       * Returns the address.
+       *
+       * @returns A {@link https://github.com/feross/buffer|Buffer}  representation of the address
+       */
+    getAddress:() => Buffer;
+
+    /**
+     * Returns the address's string representation.
+     * 
+     * @returns A string representation of the address
+     */
+    getAddressString:() => string
+  
+    constructor() {}
+  }
+  
+  /**
+   * Class for representing a key chain in Avalanche.
+   * All endpoints that need key chains should extend on this class.
+   *
+   * @typeparam KPClass extending [[KeyPair]] which is used as the key in [[KeyChain]]
+   */
+  export class KeyChain<KPClass extends KeyPair> {
+    protected keys:{[address: string]: KPClass} = {};
+  
+    /**
+       * Makes a new [[KeyPair]], returns the address.
+       *
+       * @returns Address of the new [[KeyPair]]
+       */
+    makeKey:() => Buffer;
+  
+    /**
+       * Given a private key, makes a new [[KeyPair]], returns the address.
+       *
+       * @param privk A {@link https://github.com/feross/buffer|Buffer} representing the private key
+       *
+       * @returns Address of the new [[KeyPair]]
+       */
+    importKey:(privk:Buffer) => Buffer;
+  
+    /**
+       * Gets an array of addresses stored in the [[KeyChain]].
+       *
+       * @returns An array of {@link https://github.com/feross/buffer|Buffer}  representations
+       * of the addresses
+       */
+    getAddresses = ():Array<Buffer> => Object.values(this.keys).map((kp) => kp.getAddress());
+  
+    /**
+       * Gets an array of addresses stored in the [[KeyChain]].
+       *
+       * @returns An array of string representations of the addresses
+       */
+    getAddressStrings = ():Array<string> => Object.values(this.keys)
+      .map((kp) => kp.getAddressString());
+  
+    /**
+       * Adds the key pair to the list of the keys managed in the [[KeyChain]].
+       *
+       * @param newKey A key pair of the appropriate class to be added to the [[KeyChain]]
+       */
+    addKey = (newKey:KPClass) => {
+      this.keys[newKey.getAddress().toString('hex')] = newKey;
+    };
+  
+    /**
+       * Removes the key pair from the list of they keys managed in the [[KeyChain]].
+       *
+       * @param key A {@link https://github.com/feross/buffer|Buffer} for the address or
+       * KPClass to remove
+       *
+       * @returns The boolean true if a key was removed.
+       */
+    removeKey = (key:KPClass | Buffer) => {
+      let kaddr:string;
+      if (key instanceof Buffer) {
+        kaddr = key.toString('hex');
+      } else {
+        kaddr = key.getAddress().toString('hex');
+      }
+      if (kaddr in this.keys) {
+        delete this.keys[kaddr];
+        return true;
+      }
+      return false;
+    };
+  
+    /**
+       * Checks if there is a key associated with the provided address.
+       *
+       * @param address The address to check for existence in the keys database
+       *
+       * @returns True on success, false if not found
+       */
+    hasKey = (address:Buffer):boolean => (address.toString('hex') in this.keys);
+  
+    /**
+       * Returns the [[KeyPair]] listed under the provided address
+       *
+       * @param address The {@link https://github.com/feross/buffer|Buffer} of the address to
+       * retrieve from the keys database
+       *
+       * @returns A reference to the [[KeyPair]] in the keys database
+       */
+    getKey = (address:Buffer): KPClass => this.keys[address.toString('hex')];
+  
+    /**
+       * Returns instance of [[KeyChain]].
+       * 
+       */
+    constructor() {}
+  }
